requirements.txt
Flask==1.1.4
gunicorn==20.1.0
PyYAML==5.4.1
scikit-learn==0.20.4
joblib==0.13.2
numpy==1.16.6
scipy==1.2.3
pandas==0.24.2
jieba==0.42.1
